# openapi.yaml
openapi: 3.0.0
info:
  title: api-freshman-yr
  version: 0.0.1
  description: A crowdsourced micro-funding platform for financing undergraduate tuition expenses.
  contact:
    name: Sean Travis Taylor
    email: engineering@freshmanyr.io
  termsOfService: api.freshmanyr.io/tos.md
...
servers:
  - url: "http://localhost:3000"
    description: development
paths:
...
  /students:
    get:
      summary: get_students
      tags:
        - students
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: {}
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties: {}
      operationId: get_students
      description: Get all students registered on the platform
      parameters:
        - schema:
            type: string
          in: header
          name: Authorization
          description: JWT Authorization
          required: true
...
  /status:
    get:
      summary: get_system_status
      tags:
        - system
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: {}
        "503":
          description: Service Unavailable
      operationId: get-status
      description: Get system status
...
  /sponsors:
    get:
      summary: get_sponsors
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: {}
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties: {}
      operationId: get_sponsors
      tags:
        - sponsors
      description: Get all sponsors registered on the platform
      parameters:
        - schema:
            type: string
          in: header
          name: Authorization
          description: JWT Authorization
          required: true
    post:
      summary: create_sponsor
      operationId: create_sponsor
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: {}
        "201":
          description: Created
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties: {}
      description: Register a new sponsor on the platform
      tags:
        - sponsors
      parameters:
        - schema:
            type: string
          in: header
          name: Authorization
          description: JWT Authorization
          required: true
...